refactor: 확장을 위한 디렉토리 및 코드 구조 변경
