+++ package.json
@@ -18,6 +18,7 @@
     "@types/mdx": "^2.0.13",
     "apexcharts": "^3.54.1",
     "esbuild": "^0.24.0",
+    "firebase": "^11.0.2",
     "flowbite": "^2.5.2",
     "gray-matter": "^4.0.3",
     "mdx-bundler": "^10.0.3",
