+++ src/app/components/main_frame/TagBlock.tsx
@@ -1,4 +1,4 @@
-import {countUsedTags} from "@/app/lib/ServerPosts";
+import {countUsedTags} from "@/app/lib/post/ServerPosts";
 import {MainContainerGrid} from "@_components/main_frame/MainContainer";
 
 export default async function TagBlock() {
