+++ src/app/page.tsx
@@ -1,11 +1,12 @@
 import TagBlock from "@/app/components/main_frame/TagBlock";
 import dynamic from "next/dynamic";
-import {getPostListData} from "@/app/lib/ServerPosts";
+import {getPostListData} from "@/app/lib/post/ServerPosts";
 import {ChartDataProps, generationChartData} from "@/app/components/main_frame/LanguageBlockData";
 import MainContainer, {MainContainerGrid} from "@_components/main_frame/MainContainer";
 import RecencyPostsBlock from "@_components/main_frame/RecencyPostsBlock";
-import {generationPostCardProps, Paging} from "@/app/lib/ClientPost";
+import {generationPostCardProps} from "@/app/lib/post/ClientPost";
 import Banner from "@_components/main_frame/Banner";
+import {Paging} from "@/app/lib/post/PostConfig";
 
 export async function RecencyPostBlockData() {
     const allList = [...await getPostListData('/main'), ...await getPostListData('/proj')];
