+++ src/app/posts/list/page.tsx
@@ -1,7 +1,8 @@
-import {getPostListData} from "@/app/lib/ServerPosts";
+import {getPostListData} from "@/app/lib/post/ServerPosts";
 import MainContainer from "@_components/main_frame/MainContainer";
 import MainPostListRender from "@_components/post/main/MainPostListRender";
-import {Paging} from "@/app/lib/ClientPost";
+
+import {Paging} from "@/app/lib/post/PostConfig";
 
 
 export default async function Page() {
